+++ libft/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: tsomacha <tsomacha@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 16:23:34 by tsomacha          #+#    #+#             */
-/*   Updated: 2024/11/04 16:24:26 by tsomacha         ###   ########.fr       */
+/*   Updated: 2024/11/06 16:53:47 by tsomacha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
-	unsigned int	count;
+	size_t	count;
 
 	count = 0;
 	while (*(str + count))
